Extract parsed_fragment
